+++ src/components/Tokens/TokenDetails/Resource.tsx
