+++ src/components/Tokens/TokenDetails/ShareButton.tsx
