+++ src/components/Tokens/TokenTable/FavoriteButton.tsx
