+++ src/components/Tokens/TokenTable/SearchBar.tsx
