+++ src/theme/components.tsx
@@ -1,6 +1,6 @@
 import { Trans } from '@lingui/macro'
 import { outboundLink } from 'components/analytics'
-import { MOBILE_MEDIA_BREAKPOINT } from 'components/Explore/constants'
+import { MOBILE_MEDIA_BREAKPOINT } from 'components/Tokens/constants'
 import useCopyClipboard from 'hooks/useCopyClipboard'
 import React, { forwardRef, HTMLProps, ReactNode, useCallback, useImperativeHandle } from 'react'
 import {
